상인 개선
아이템을 구매할때마다 가격이 오름

+++ Assets/Junhui/Script/GameManager.cs
@@ -30,7 +30,7 @@ public class GameManager : MonoBehaviour
 
     public float hp = 100f;
     public float maxHp = 100f;
-    public int gold = 0;
+    public float gold = 0;
     public float damage = 1;
 
     public int unlockedMapCount = 2;
